nml: convert arch-dependent types to fixed-width types
.. when they are related to types that pass over the wire in nml.

include stdint.h as needed.

use inttypes.h integer formatting string constants as appropriate.

+++ src/emc/nml_intf/emc.cc
@@ -476,7 +476,7 @@ int emcFormat(NMLTYPE type, void *buffer, CMS * cms)
 }
 
 // NML Symbol Lookup Function
-const char *emc_symbol_lookup(long type)
+const char *emc_symbol_lookup(uint32_t type)
 {
     switch (type) {
     case EMC_ABORT_TYPE:
